Move tower saving/reloading into consensus
